Automates releases with GitHub Actions (#4)
TL;DR
------

Adds CD capabilities that produce a GitHub release and push the bundle to GitHub Container Registry

Details
-------

Continuously delivers the polices to a GitHub release and an OPA Bundle in the GitHub Container Registry whenever a new tag marking a release is applied to `main`.

The workflow:

* Runs the policy tests with `conftest verify`
* Uses a custom action that creates an OPA bundle with the `opa` CLI and releases that bundle on GitHub
* Pushes the release to the GitHub Container Registry so `conftest` uses can invoke it directly from there

At this time there are two distinct bundles: one that is available as a GitHub release (created with `opa` build) and one that
is created by Conftest when pushed to the registry with `conftest push`. This was necessary because current tooling makes it difficult to push a pre-created bundle to an OCI registry. I suspect that situation to remedy itself.