Updates workflow to include a version tag
TL;DR
-----

Allows older versions to be pulled by pushing with a version tag

Details
-------

Updates the Conftest push action to specify the tag variable, then
uses the updated action to tag the bundle with the version tag.
Also continues to push it with the `latest` tag so that the most
recent version is available by pulling `latest`.