Explicity sets the release note to the head commit message (#11)
TL;DR
-----

Sets the release note to the commit message on the head

Details
-------

Explicitly sets the release note when cutting the release. Uses
the commit message from the head commit, which it will show with
the release if there isn't a release note. Setting it explicitly
improves the formatting when looking at the list of releases.